fix(poke): keep GCS roles storage out of SPA bundle
